Move stuff to utils file
