+++ web/src/app/admin/configuration/llm/interfaces.ts
@@ -1,16 +1,3 @@
-import {
-  AnthropicIcon,
-  AmazonIcon,
-  CPUIcon,
-  MicrosoftIconSVG,
-  MistralIcon,
-  MetaIcon,
-  GeminiIcon,
-  IconProps,
-  DeepseekIcon,
-  OpenAISVG,
-} from "@/components/icons/icons";
-
 export interface CustomConfigKey {
   name: string;
   display_name: string;
@@ -88,45 +75,3 @@ export interface LLMProviderDescriptor {
   groups: number[];
   model_configurations: ModelConfiguration[];
 }
-
-export const getProviderIcon = (providerName: string, modelName?: string) => {
-  const iconMap: Record<
-    string,
-    ({ size, className }: IconProps) => JSX.Element
-  > = {
-    amazon: AmazonIcon,
-    phi: MicrosoftIconSVG,
-    mistral: MistralIcon,
-    ministral: MistralIcon,
-    llama: MetaIcon,
-    gemini: GeminiIcon,
-    deepseek: DeepseekIcon,
-    claude: AnthropicIcon,
-    anthropic: AnthropicIcon,
-    openai: OpenAISVG,
-    microsoft: MicrosoftIconSVG,
-    meta: MetaIcon,
-    google: GeminiIcon,
-  };
-
-  // First check if provider name directly matches an icon
-  if (providerName.toLowerCase() in iconMap) {
-    return iconMap[providerName.toLowerCase()];
-  }
-
-  // Then check if model name contains any of the keys
-  if (modelName) {
-    const lowerModelName = modelName.toLowerCase();
-    for (const [key, icon] of Object.entries(iconMap)) {
-      if (lowerModelName.includes(key)) {
-        return icon;
-      }
-    }
-  }
-
-  // Fallback to CPU icon if no matches
-  return CPUIcon;
-};
-
-export const isAnthropic = (provider: string, modelName: string) =>
-  provider === "anthropic" || modelName.toLowerCase().includes("claude");
